+++ sampleproject/components/recursive.py
@@ -1,3 +1,4 @@
+import time
 from typing import Any, Dict
 
 from django_components import Component, register, types
@@ -5,17 +6,18 @@ from django_components import Component, register, types
 
 @register("recursive")
 class Recursive(Component):
-    def get(self, request):
-        import time
-        time_before = time.time()
-        output = self.render_to_response(
-            kwargs={
-                "depth": 0,
-            },
-        )
-        time_after = time.time()
-        print("TIME: ", time_after - time_before)
-        return output
+    class View:
+        def get(self, request):
+            time_before = time.time()
+            output = Recursive.render_to_response(
+                request=request,
+                kwargs={
+                    "depth": 0,
+                },
+            )
+            time_after = time.time()
+            print("TIME: ", time_after - time_before)
+            return output
 
     def get_context_data(self, depth: int = 0) -> Dict[str, Any]:
         return {"depth": depth + 1}