+++ src/django_components/app_settings.py
@@ -751,9 +751,10 @@ class InternalSettings:
 
         # Prepend built-in extensions
         from django_components.extensions.defaults import DefaultsExtension
+        from django_components.extensions.url import UrlExtension
         from django_components.extensions.view import ViewExtension
 
-        extensions = [DefaultsExtension, ViewExtension] + list(extensions)
+        extensions = [DefaultsExtension, ViewExtension, UrlExtension] + list(extensions)
 
         # Extensions may be passed in either as classes or import strings.
         extension_instances: List["ComponentExtension"] = []