+++ src/django_components/extensions/view.py
@@ -15,8 +15,19 @@ class ViewFn(Protocol):
 
 class ComponentView(ComponentExtension.ExtensionClass, View):  # type: ignore
     """
-    Subclass of `django.views.View` where the `Component` instance is available
+    The interface for `Component.View`.
+
+    Override the methods of this class to define the behavior of the component.
+
+    This class is a subclass of `django.views.View`. The `Component` instance is available
     via `self.component`.
+
+    **Example:**
+    ```python
+    class MyComponent(Component):
+        class View:
+            def get(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
+                return HttpResponse("Hello, world!")
     """
 
     # NOTE: This attribute must be declared on the class for `View.as_view()` to allow
@@ -33,42 +44,40 @@ class ComponentView(ComponentExtension.ExtensionClass, View):  # type: ignore
     # Each method actually delegates to the component's method of the same name.
     # E.g. When `get()` is called, it delegates to `component.get()`.
 
+    # TODO_V1 - In v1 handlers like `get()` should be defined on the Component.View class,
+    #           not the Component class directly. This is to align Views with the extensions API
+    #           where each extension should keep its methods in the extension class.
+    #           Instead, the defaults for these methods should be something like
+    #           `return self.component.render_to_response()` or similar.
     def get(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "get")(request, *args, **kwargs)
+        return getattr(self.component, "get")(request, *args, **kwargs)
 
     def post(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "post")(request, *args, **kwargs)
+        return getattr(self.component, "post")(request, *args, **kwargs)
 
     def put(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "put")(request, *args, **kwargs)
+        return getattr(self.component, "put")(request, *args, **kwargs)
 
     def patch(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "patch")(request, *args, **kwargs)
+        return getattr(self.component, "patch")(request, *args, **kwargs)
 
     def delete(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "delete")(request, *args, **kwargs)
+        return getattr(self.component, "delete")(request, *args, **kwargs)
 
     def head(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "head")(request, *args, **kwargs)
+        return getattr(self.component, "head")(request, *args, **kwargs)
 
     def options(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "options")(request, *args, **kwargs)
+        return getattr(self.component, "options")(request, *args, **kwargs)
 
     def trace(self, request: HttpRequest, *args: Any, **kwargs: Any) -> HttpResponse:
-        component: "Component" = self.component
-        return getattr(component, "trace")(request, *args, **kwargs)
+        return getattr(self.component, "trace")(request, *args, **kwargs)
 
 
 class ViewExtension(ComponentExtension):
     """
     This extension adds a nested `View` class to each `Component`.
+
     This nested class is a subclass of `django.views.View`, and allows the component
     to be used as a view by calling `ComponentView.as_view()`.