+++ tests/components/multi_file/multi_file.py
@@ -9,12 +9,19 @@ from django_components import Component, register
 class MultFileComponent(Component):
     template_file = "multi_file/multi_file.html"
 
-    def post(self, request, *args, **kwargs) -> HttpResponse:
-        variable = request.POST.get("variable")
-        return self.render_to_response({"variable": variable})
+    class View:
+        def post(self, request, *args, **kwargs) -> HttpResponse:
+            variable = request.POST.get("variable")
+            return MultFileComponent.render_to_response(
+                request=request,
+                kwargs={"variable": variable},
+            )
 
-    def get(self, request, *args, **kwargs) -> HttpResponse:
-        return self.render_to_response({"variable": "GET"})
+        def get(self, request, *args, **kwargs) -> HttpResponse:
+            return MultFileComponent.render_to_response(
+                request=request,
+                kwargs={"variable": "GET"},
+            )
 
     def get_context_data(self, variable, *args, **kwargs) -> Dict[str, Any]:
         return {"variable": variable}