+++ tests/components/relative_file/relative_file.py
@@ -13,12 +13,19 @@ class RelativeFileComponent(Component):
         js = "relative_file.js"
         css = "relative_file.css"
 
-    def post(self, request, *args, **kwargs) -> HttpResponse:
-        variable = request.POST.get("variable")
-        return self.render_to_response({"variable": variable})
+    class View:
+        def post(self, request, *args, **kwargs) -> HttpResponse:
+            variable = request.POST.get("variable")
+            return RelativeFileComponent.render_to_response(
+                request=request,
+                kwargs={"variable": variable},
+            )
 
-    def get(self, request, *args, **kwargs) -> HttpResponse:
-        return self.render_to_response({"variable": "GET"})
+        def get(self, request, *args, **kwargs) -> HttpResponse:
+            return RelativeFileComponent.render_to_response(
+                request=request,
+                kwargs={"variable": "GET"},
+            )
 
     def get_context_data(self, variable, *args, **kwargs) -> Dict[str, Any]:
         return {"variable": variable}