+++ tests/test_extension.py
@@ -22,6 +22,7 @@ from django_components.extension import (
 )
 from django_components.extensions.defaults import DefaultsExtension
 from django_components.extensions.view import ViewExtension
+from django_components.extensions.url import UrlExtension
 
 from django_components.testing import djc_test
 from .testutils import setup_test_config
@@ -126,10 +127,11 @@ def with_registry(on_created: Callable):
 class TestExtension:
     @djc_test(components_settings={"extensions": [DummyExtension]})
     def test_extensions_setting(self):
-        assert len(app_settings.EXTENSIONS) == 3
+        assert len(app_settings.EXTENSIONS) == 4
         assert isinstance(app_settings.EXTENSIONS[0], DefaultsExtension)
         assert isinstance(app_settings.EXTENSIONS[1], ViewExtension)
-        assert isinstance(app_settings.EXTENSIONS[2], DummyExtension)
+        assert isinstance(app_settings.EXTENSIONS[2], UrlExtension)
+        assert isinstance(app_settings.EXTENSIONS[3], DummyExtension)
 
     @djc_test(components_settings={"extensions": [DummyExtension]})
     def test_access_component_from_extension(self):
@@ -152,7 +154,7 @@ class TestExtension:
 class TestExtensionHooks:
     @djc_test(components_settings={"extensions": [DummyExtension]})
     def test_component_class_lifecycle_hooks(self):
-        extension = cast(DummyExtension, app_settings.EXTENSIONS[2])
+        extension = cast(DummyExtension, app_settings.EXTENSIONS[3])
 
         assert len(extension.calls["on_component_class_created"]) == 0
         assert len(extension.calls["on_component_class_deleted"]) == 0
@@ -184,7 +186,7 @@ class TestExtensionHooks:
 
     @djc_test(components_settings={"extensions": [DummyExtension]})
     def test_registry_lifecycle_hooks(self):
-        extension = cast(DummyExtension, app_settings.EXTENSIONS[2])
+        extension = cast(DummyExtension, app_settings.EXTENSIONS[3])
 
         assert len(extension.calls["on_registry_created"]) == 0
         assert len(extension.calls["on_registry_deleted"]) == 0
@@ -221,7 +223,7 @@ class TestExtensionHooks:
                 return {"name": name}
 
         registry.register("test_comp", TestComponent)
-        extension = cast(DummyExtension, app_settings.EXTENSIONS[2])
+        extension = cast(DummyExtension, app_settings.EXTENSIONS[3])
 
         # Verify on_component_registered was called
         assert len(extension.calls["on_component_registered"]) == 1
@@ -259,7 +261,7 @@ class TestExtensionHooks:
         test_slots = {"content": "Some content"}
         TestComponent.render(context=test_context, args=("arg1", "arg2"), kwargs={"name": "Test"}, slots=test_slots)
 
-        extension = cast(DummyExtension, app_settings.EXTENSIONS[2])
+        extension = cast(DummyExtension, app_settings.EXTENSIONS[3])
 
         # Verify on_component_input was called with correct args
         assert len(extension.calls["on_component_input"]) == 1